manage movies page stuff and a small thing about the admin movie card
There was a div which had an attribute called "class" which isn't allowed in React; React uses "className" so it doesn't get mixed up with the "class" keyword that JavaScript has. As for the forms for entering movie info, there is no easy way to get the info from the form fields, so using useState and updating methods is a kind of work around.